requirements.txt
anyio==3.6.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
fastapi-utils==0.2.1
google-auth==2.8.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
kubernetes==24.2.0
oauthlib==3.2.0
prometheus-client==0.14.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
starlette-exporter==0.13.0
typing-extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websocket-client==1.3.3
websockets==10.3
